requirements.txt
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-html>=4.0.0
pyyaml>=6.0
jinja2>=3.1.0
allure-pytest>=2.13.0
click>=8.1.0
pydantic>=2.0.0
